commit coy

+++ application/controllers/master/jam_kerja.php
@@ -50,7 +50,7 @@ class Jam_kerja extends CI_Controller {
         $workday_path = addslashes($_POST['workday_path']);
         $filename = "assets/schedules/" . $workday_path;
        
-        if($workday_path != ''){
+        if($workday_path != '' && file_exists($filename)){
              unlink($filename);
         }
         if($this->record->delete($workday_id))
@@ -64,42 +64,24 @@ class Jam_kerja extends CI_Controller {
     {
         if(!isset($_POST))	
             show_404();
-
-        $path = $_FILES["workday_path"]["name"];
         
-        if($this->record->cekPath($path)){
+        $filename = "assets/schedules/" . $_FILES["workday_path"]["name"];
+        if(file_exists($filename)){
             echo json_encode(array('ada'=>true));
-        } 
-        if($this->record->upload($workday_id)) {
-            move_uploaded_file($_FILES["workday_path"]["tmp_name"],
-            "assets/schedules/" . $_FILES["workday_path"]["name"]);
-            echo json_encode(array('success'=>true));
-        } else {
-            echo json_encode(array('msg'=>'Gagal mengubah data'));
+        }else {            
+            if($this->record->cekPath($workday_id)){
+                $this->record->deleteFile($workday_id);
+            }
+            if($this->record->upload($workday_id)) {
+                move_uploaded_file($_FILES["workday_path"]["tmp_name"],
+                "assets/schedules/" . $_FILES["workday_path"]["name"]);
+                echo json_encode(array('success'=>true));
+            } else {
+                echo json_encode(array('msg'=>'Gagal mengubah data'));
+            }
         }
-            
     }
     
-    
-    public function deleteFile()
-    {
-        if(!isset($_POST))	
-            show_404();
-       
-        $workday_id = $this->input->post('workday_id',true);
-        $workday_path = $this->input->post('workday_path',true);
-        $filename = "assets/schedules/" . $workday_path;
-        
-        if($workday_path != '' && file_exists($filename)){
-             unlink($filename);
-        }
-        if($this->record->deleteFile($workday_id))
-            echo json_encode(array('success'=>true));
-        else
-            echo json_encode(array('msg'=>'Gagal mengubah data'));
-                
-    }
-               
 }
 
 /* End of file welcome.php */
